Rename pyanaconda.payload.__init__
Move the classes Payload and PackagePayload to pyanaconda.payload.base,
but keep their history. We will create a new __init__ file in the following
commit to achieve that.
