+++ pyanaconda/payload/base.py
